requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
PyJWT==2.8.0
Werkzeug==2.3.7
flask-swagger-ui==4.11.1
apispec==6.3.0
marshmallow==3.20.1
Authlib>=1.2.0
cryptography>=41.0.0
python-jose>=3.3.0